+++ include/UECS/SystemMngr.h
@@ -11,10 +11,13 @@ namespace Ubpa::UECS {
 	public:
 		SystemTraits systemTraits;
 		
-		SystemMngr(World* w) : w{w}{}
+		SystemMngr(World* w) : w{w} {}
 		SystemMngr(const SystemMngr& mngr, World* w);
 		SystemMngr(SystemMngr&& mngr, World* w) noexcept;
 		~SystemMngr();
+
+		const auto& GetAliveSystemIDs() const noexcept { return aliveSystemIDs; }
+		const auto& GetActiveSystemsIDs() const noexcept { return activeSystemIDs; }
 
 		// not alive -> create
 		void Create(size_t systemID);
@@ -33,8 +36,32 @@ namespace Ubpa::UECS {
 		bool IsAlive(size_t systemID) const;
 		bool IsActive(size_t systemID) const;
 
-		const auto& GetAliveSystemIDs() const noexcept { return aliveSystemIDs; }
-		const auto& GetActiveSystemsIDs() const noexcept { return activeSystemIDs; }
+		// [ Template ] Functions
+		///////////////////////////
+
+		template<typename... Systems>
+		void Create() { (Create(systemTraits.GetID<Systems>()), ...); }
+
+		template<typename... Systems>
+		void Activate() { (Activate(systemTraits.GetID<Systems>()), ...); }
+
+		template<typename... Systems>
+		void Deactivate() { (Deactivate(systemTraits.GetID<Systems>()), ...); }
+
+		template<typename... Systems>
+		void Destroy() { (Destroy(systemTraits.GetID<Systems>()), ...); }
+
+		template<typename System>
+		bool IsAlive() const { return IsAlive(systemTraits.GetID<System>()); }
+
+		template<typename System>
+		bool IsActive() const { return IsActive(systemTraits.GetID<System>()); }
+
+		template<typename... Systems>
+		std::array<size_t, sizeof...(Systems)> RegisterAndCreate();
+
+		template<typename... Systems>
+		std::array<size_t, sizeof...(Systems)> RegisterAndActivate();
 
 		SystemMngr(const SystemMngr&) = delete;
 		SystemMngr(SystemMngr&&) noexcept = delete;
@@ -51,3 +78,5 @@ namespace Ubpa::UECS {
 		std::unordered_set<size_t> activeSystemIDs;
 	};
 }
+
+#include "detail/SystemMngr.inl"