+++ include/UECS/SystemTraits.h
@@ -37,32 +37,51 @@ namespace Ubpa::UECS {
 		SystemTraits() = default;
 		SystemTraits(const SystemTraits&);
 		SystemTraits(SystemTraits&&) noexcept = default;
+		SystemTraits& operator=(SystemTraits&);
+		SystemTraits& operator=(SystemTraits&&) noexcept = default;
 
+		// register system's name and get an ID
+		// if it is already registered, return it's ID directly
 		size_t Register(std::string name);
-		void RegisterOnCreate(size_t ID, std::function<OnCreate>);
-		void RegisterOnActivate(size_t ID, std::function<OnActivate>);
-		void RegisterOnUpdate(size_t ID, std::function<OnUpdate>);
-		void RegisterOnDeactivate(size_t ID, std::function<OnDeactivate>);
-		void RegisterOnDestroy(size_t ID, std::function<OnDestroy>);
 
-		bool IsRegistered(size_t ID) const noexcept;
-		size_t GetID(std::string_view name) const;
+		// ID must exist
+		void RegisterOnCreate    (size_t ID, std::function<OnCreate>);
+		void RegisterOnActivate  (size_t ID, std::function<OnActivate>);
+		void RegisterOnUpdate    (size_t ID, std::function<OnUpdate>);
+		void RegisterOnDeactivate(size_t ID, std::function<OnDeactivate>);
+		void RegisterOnDestroy   (size_t ID, std::function<OnDestroy>);
+
 		std::string_view Nameof(size_t ID) const noexcept;
+		size_t GetID(std::string_view name) const;
+		bool IsRegistered(size_t ID) const noexcept;
 		const auto& GetNameIDMap() const noexcept { return name2id; }
 
-		template<typename... Systems>
-		std::array<size_t, sizeof...(Systems)> Register();
+		// [ Template ] functions
+		///////////////////////////
+
 		template<typename System>
 		static std::string_view StaticNameof() noexcept;
+
+		// for each <System> in <Systems...>
+		// 1. Register(StaticNameof<System>())
+		// 2. RegisterOn{Create|Activate|Update|Deactivate|Destroy} if <System> has them
+		template<typename... Systems>
+		std::array<size_t, sizeof...(Systems)> Register();
+
+		template<typename System>
+		size_t GetID() const { return GetID(StaticNameof<System>()); }
+
+		template<typename System>
+		bool IsRegistered() const { return GetID<System>(); }
 		
 	private:
 		friend class SystemMngr;
 		
-		void Create(size_t ID, World*) const;
-		void Activate(size_t ID, World*) const;
-		void Update(size_t ID, Schedule&) const;
+		void Create    (size_t ID, World*) const;
+		void Activate  (size_t ID, World*) const;
+		void Update    (size_t ID, Schedule&) const;
 		void Deactivate(size_t ID, World*) const;
-		void Destroy(size_t ID, World*) const;
+		void Destroy   (size_t ID, World*) const;
 
 		std::vector<std::string> names;
 		std::unordered_map<std::string_view, size_t> name2id;