+++ src/test/00_basic/main.cpp
@@ -18,8 +18,7 @@ struct MoverSystem {
 
 int main() {
 	World w;
-	auto [move] = w.systemMngr.systemTraits.Register<MoverSystem>();
+	w.systemMngr.RegisterAndActivate<MoverSystem>();
 	w.entityMngr.Create<Position, Velocity>();
-	w.systemMngr.Activate(move);
 	w.Update();
 }