+++ src/test/02_order/main.cpp
@@ -23,7 +23,7 @@ struct DataSystem {
 
 int main() {
 	World w;
-	auto [dataSystem] = w.systemMngr.systemTraits.Register<DataSystem>();
+	w.systemMngr.RegisterAndActivate<DataSystem>();
 
 	w.entityMngr.cmptTraits.Register<
 		Data1,
@@ -32,7 +32,6 @@ int main() {
 
 	w.entityMngr.Create<Data1, Data2>();
 
-	w.systemMngr.Activate(dataSystem);
 	w.Update();
 
 	cout << w.DumpUpdateJobGraph() << endl;