+++ src/test/07_overload/main.cpp
@@ -24,9 +24,7 @@ struct AVP_System {
 
 int main() {
 	World w;
-	auto systemIndices = w.systemMngr.systemTraits.Register<VP_System, AVP_System>();
-	for (auto idx : systemIndices)
-		w.systemMngr.Activate(idx);
+	w.systemMngr.RegisterAndActivate<VP_System, AVP_System>();
 
 	w.entityMngr.Create<V, P>();
 	w.entityMngr.Create<A, V, P>();